Audit: do not print error when LSMs disabled
RHBZ: 785936

If the audit system collects a record about one process sending a signal
to another process it includes in that collection the 'secid' or 'an int
used to represet an LSM label.'  If there is no LSM enabled it will
collect a 0.  The problem is that when we attempt to print that record
we ask the LSM to convert the secid back to a string.  Since there is no
LSM it returns EOPNOTSUPP.

Most code in the audit system checks if the secid is 0 and does not
print LSM info in that case.  The signal information code however forgot
that check.  Thus users will see a message in syslog indicating that
converting the sid to string failed.  Add the right check.

Signed-off-by: Eric Paris <eparis@redhat.com>
(cherry picked from commit ad395abece974e50cfd7ddd509a4faae8e238a40)

Orabug: 17537769

Signed-off-by: Jerry Snitselaar <jerry.snitselaar@oracle.com>
Acked-by: Dan Duval <dan.duval@oracle.com>

(cherry picked from commit 274fb72160e7070cc7923ceeb5bba1087e0bdcf6)
Signed-off-by: Jerry Snitselaar <jerry.snitselaar@oracle.com>