core/xen: Use _PAGE_IOMAP when mmaping PCI BARs.
Xen PV guests (which the initial domain - dom0 is) has a lookup
mechanism called Page Frame Number (PFN) to Machine Frame Number (MFN)
where it can find the proper bus address for DMA operations.

Most of the time the PFNs != MFNs (this is after all virtualization)
and that is OK. For DMA operations we need the MFN value as that is
what we pass in the PCI devices and they need to use the correct
bus address. As such PV guests have this P2M lookup mechanism (a tree)
to figure out what the MFN is. That is all good for all memory
regions that are enumerated in the E820 table - we can determine that
the E820_RSV and the "gaps" (aka MMIO regions) should return an
identity mapping - ie, pfn == mfn. As such we mark them in the P2M.

Unfortunatly not all information is present in the E820. For the
newer type of InfiniBand cards the MMIO regions are outside of the
E820. Even worst, they are outside the P2M tree (which can only
hold up to 500MB of PFN lookups). Any entry outside the P2M
tree returns a MFN of zero.  And the Infiniband controllers are in:

pci_root PNP0A08:00: host bridge window [mem 0x80000000-0xc7ffffff]
pci_root PNP0A08:00: host bridge window [mem 0x3c0000000000-0x3c0fffffffff] <==

which is at 60TB range. Way out of the P2M tree.

The solution to this problem is that we can:
 a) Assume any entries outside the P2M should return identity mapping
   (PFN == MFN).
 b) Stick _PAGE_IOMAP when mapping as that will automatically assume
   PFN == MFN.
 c) Stick _PAGE_IOMAP on all calls to io_remap_pfn_range.

The a) is being persued upstream - but it requires a dilligient
look at all the P2M code as the default assumption is that any "unknown"
region is INVALID_P2M_ENTRY (-1UL) - and migration code checks for that.

c) looks tempting and it has no adverse effect on baremetal (as _PAGE_IOMAP
on baremetal is just an extra bit in the PTE that is not used), while
b) can be done just in the drivers.

This patch does the change in the 'c)' path.

P.S.
c) surfaced some time upstream and the x86 maintainers weren't too
happy with it. They recommended that the P2M code handle this correctly.
Which it does for the normal case when a PCI device is nicely
enumerated within the 500MB range.

Orabug: 17753790

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Acked-by: Jerry Snitselaar <jerry.snitselaar@oracle.com>